// snl-gms-mock/packages/common-model/src/ts/common/types.ts
export enum Units {
  DEGREES = 'DEGREES',
  DECIBELS = 'DECIBELS',
  RADIANS = 'RADIANS',
  SECONDS = 'SECONDS',
  HERTZ = 'HERTZ',
  DEGREES_PER_SECOND = 'DEGREES_PER_SECOND',
  SECONDS_PER_DEGREE = 'SECONDS_PER_DEGREE',
  SECONDS_PER_RADIAN = 'SECONDS_PER_RADIAN',
  SECONDS_PER_DEGREE_SQUARED = 'SECONDS_PER_DEGREE_SQUARED',
  SECONDS_PER_KILOMETER_SQUARED = 'SECONDS_PER_KILOMETER_SQUARED',
  SECONDS_PER_KILOMETER = 'SECONDS_PER_KILOMETER',
  SECONDS_PER_KILOMETER_PER_DEGREE = 'SECONDS_PER_KILOMETER_PER_DEGREE',
  ONE_OVER_KM = 'ONE_OVER_KM',
  NANOMETERS = 'NANOMETERS',
  NANOMETERS_PER_SECOND = 'NANOMETERS_PER_SECOND',
  NANOMETERS_PER_COUNT = 'NANOMETERS_PER_COUNT',
  NANOMETERS_SQUARED_PER_SECOND = 'NANOMETERS_SQUARED_PER_SECOND',
  UNITLESS = 'UNITLESS',
  MAGNITUDE = 'MAGNITUDE',
  COUNTS_PER_NANOMETER = 'COUNTS_PER_NANOMETER',
  COUNTS_PER_PASCAL = 'COUNTS_PER_PASCAL',
  PASCALS_PER_COUNT = 'PASCALS_PER_COUNT',
  ONE_OVER_DEGREE = 'ONE_OVER_DEGREE',
  SECONDS_PER_DEGREE_KM = 'SECONDS_PER_DEGREE_KM',
  DEGREES_PER_KM = 'DEGREES_PER_KM'
}