// snl-gms-mock/packages/common-model/src/ts/fk/types.ts
export interface FkFrequencyPreview {
  readonly frequencyBand: FkFrequencyRange;
  readonly fkSpectra: FkSpectraClaimCheck;
}
// ...
export interface FkWindow {
  duration: number;
  lead: number;
}
export interface FkSpectraMetadata {
  readonly fkSpectrumWindow: FkWindow;
  readonly slownessGrid: SlownessGrid;
  readonly phase: string;
}
/**
 * Fk spectra COI representation
 */
export interface FkSpectraCOI extends Timeseries {
  readonly fkSpectraMetadata?: FkSpectraMetadata;
  readonly samples: FkSpectrum[];
}

/**
 * Fk spectra with UI representation
 */
export interface FkSpectraClaimCheck extends TimeseriesClaimCheck {
  // Needed for UI processing added when query returns FkPowerSpectraCOI
  readonly configuration: FkSpectraTemplate;

  /** Tracking the accepted configuration, need to know on a new compute if params changed so can be reverted */
  readonly _uiAcceptedConfiguration?: FkSpectraTemplate;
}
// ...
export interface FkSpectrum {
  readonly power: number[][];
  readonly fstat: number[][];
  readonly fkQual?: number;
  readonly fkAttributes?: FkAttributes;
}
export interface FkAttributes {
  readonly peakFStat: number;
  readonly peakPower: number;
  readonly receiverToSourceAzimuth: DoubleValue;
  readonly slowness: DoubleValue;
}

/**
 * FK Spectra Template Definition Components
 */
export interface SlownessGrid {
  maxSlowness: number;
  numPoints: number;
}
export interface FkWaveformSampleRate {
  waveformSampleRateHz: number;
  waveformSampleRateToleranceHz: number;
}
export interface FkFrequencyRange {
  lowFrequencyHz: number;
  highFrequencyHz: number;
}
// ...
export interface FkSpectraParameters {
  phase: string;
  preFilter: FilterDefinition | undefined;
  slownessGrid: SlownessGrid;
  fkSpectrumWindow: FkWindow;
  fkFrequencyRange: FkFrequencyRange;
  waveformSampleRate: FkWaveformSampleRate;
  spectrumStepDuration: number;
  orientationAngleToleranceDeg: number;
  minimumWaveformsForSpectra: number;
  normalizeWaveforms: boolean;
  twoDimensional: boolean;
  fftTaperFunction: TaperFunction;
  fkUncertaintyOption: FkUncertaintyOption;
  fftTaperPercent: number;
}
export interface FkSpectraDefinition {
  orientationAngles: OrientationAngles;
  fkParameters: FkSpectraParameters;
}
export interface FkSpectraTemplate {
  inputChannels: EntityReference<'name', Channel>[];
  fkSpectraParameters: FkSpectraParameters;
  fkSpectraWindow: FkWindow;
  station: VersionReference<'name', Station>;
  phaseType: string;
}